Refactor Auth module
