+++ src/modules/auth/domain/auth.user.ts
